requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-modeltranslation==0.19.15
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
mccabe==0.7.0
psycopg2-binary
packaging==25.0
pillow==11.2.1
pycodestyle==2.14.0
pyflakes==3.4.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.3
python-dotenv
gunicorn
dj-database-url
whitenoise

